finalizando rest api

+++ routers/orders.js
@@ -1,9 +1,10 @@
 const express = require('express');
 const router = express.Router();
 const Order = require('../models/Orders');
-const { validateJWT, validateRoles } = require('../middlewares/validate-jwt');
+const Orderdet = require('../models/Orderdet');
 const Product = require('../models/Product');
 const User = require('../models/User');
+const { validateJWT, validateRoles } = require('../middlewares/validate-jwt');
 
 
 router.get('/ordersadmin', [validateJWT, validateRoles], async(req, res) => {
@@ -16,37 +17,59 @@ router.get('/ordersadmin', [validateJWT, validateRoles], async(req, res) => {
     });
 });
 
-
 router.post('/', [validateJWT], async(req, res) => {
-
-    const { product_id, quantity_orders } = req.body;
-    const selectProduct = await Product.findOne({ where: { id: product_id } });
 
     try {
 
-        if (!selectProduct) {
-            return res.status(401).json({
-                msg: `the id ${product_id} not exist.`
-            });
+        let otro = [];
+        let total = 0;
+
+        for (const item in req.body.items) {
+            const product_id = req.body.items[item]
+            otro.push(product_id.product_id)
         };
 
-        total = selectProduct.cost * quantity_orders;
+        for (const item in otro) {
+            const selectProduct = await Product.findOne({ where: { id: otro[item] } });
+            if (!selectProduct) {
+                return res.status(401).json({
+                    msg: `the id ${otro[item]} not exist.`
+                });
+            };
+            total = total + selectProduct.cost;
+        };
 
-        await Order.create({ quantity_orders, total, estado: 'new', userId: req.user.id, productId: selectProduct.id });
+        const created = await Order.create({
+            formpag: req.body.formpag,
+            total: total,
+            estado: 'nuevo',
+            userId: req.user.id
+        });
+
+        req.body.items.forEach(async(item) => {
+            const selectProduct = await Product.findOne({ where: { id: item.product_id } });
+            if (!selectProduct) {
+                return res.status(401).json({
+                    msg: `the id ${item.product_id} not exist.`
+                });
+            };
+
+            await Orderdet.create({
+                id_pedido: created.id,
+                productId: item.product_id,
+                quantity_orders: item.quantity_orders,
+                total: selectProduct.cost
+            });
+        });
+
+        res.json({...created.dataValues, items: req.body.items });
 
     } catch (error) {
         res.status(500).json(error);
     };
 
-    res.json({
-        msg: 'post api-controller',
-        quantity_orders,
-        total,
-        user: req.user.id,
-        Product: selectProduct
-    });
-
 });
+
 
 router.get('/', [validateJWT], async(req, res) => {
 
@@ -57,8 +80,9 @@ router.get('/', [validateJWT], async(req, res) => {
                 model: User,
                 attributes: ['name', 'email', 'phone', 'adress']
             }, {
-                model: Product,
-                attributes: ['title', 'cost']
+                model: Orderdet,
+                attributes: ['id_pedido', 'quantity_orders', 'total'],
+                include: [{ model: Product, attributes: ['title', 'cost'] }]
             }]
         });
         res.json({
@@ -76,10 +100,10 @@ router.get('/:orderId', [validateJWT, validateRoles], async(req, res) => {
     const { orderId } = req.params;
 
     try {
-        const order = await Order.findAll({
+        const order = await Order.findOne({
             where: { id: orderId },
             include: [{
-                model: Product
+                model: Orderdet
             }, {
                 model: User,
                 attributes: ['name', 'email', 'phone', 'adress']
@@ -109,6 +133,12 @@ router.put('/:orderId', [validateJWT, validateRoles], async(req, res) => {
     const { estado } = req.body;
 
     try {
+        const select = await Order.findOne({ where: { id: orderId } });
+        if (!select) {
+            return res.status(401).json({
+                msg: `the id ${orderId} not exist.`
+            });
+        };
         await Order.update({ estado }, { where: { id: orderId } });
 
         res.json({
@@ -125,7 +155,15 @@ router.delete('/:orderId', [validateJWT, validateRoles], async(req, res) => {
     const { orderId } = req.params;
 
     try {
+        const select = await Order.findOne({ where: { id: orderId } });
+        if (!select) {
+            return res.status(401).json({
+                msg: `the id ${orderId} not exist.`
+            });
+        };
+
         Order.destroy({ where: { id: orderId } });
+
     } catch (error) {
         res.status(500).json(error);
     };
